Update WCNSS_qcom_cfg.ini
This takes care of 2 issues:  wifi disconnect after about 15 min of sleep (and you can't change it in settings.global, because prima driver takes over android framework on power saving; and excessive battery draining due to a barrage of arp packets. This was tested succesfully on Xperia Z1 for a few days.